+++ application/views/error.php
@@ -1,0 +1,5 @@
+<div class="row">
+    <div class="card-body">
+        <b>An error has occured, the plane you specified does not exist</b>
+    </div>
+</div>